create workspace extension config service
